// front/src/api/models/uploadResultDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DataShare API
 * API de transfert de fichiers DataShare (US01-US10)
 * OpenAPI spec version: 1.0.0
 */

export interface UploadResultDto {
  downloadToken: string;
  downloadUrl: string;
  originalName: string;
  sizeBytes: number;
  mimeType: string;
  expiresAt: string;
  passwordProtected: boolean;
  tags: string[];
}